RTECO-1074 - Remove redundendent functions in agent namespace-2
